requirements.txt
python-telegram-bot==20.7
openai==1.3.7
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
python-dotenv==1.0.0
Flask==3.0.0
APScheduler==3.10.4
requests==2.31.0
psutil==5.9.6
bleach==6.1.0
gunicorn==21.2.0
aiohttp==3.9.1
pytz==2023.3
